pyproject.toml
[tool.poetry]
name = "newbook"
version = "0.1.0"
description = ""
authors = ["salehmmasri <salehalmasri494@yahoo.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.1"
djangorestframework = "^3.12.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
